fix(FEC-13141): New gear - Captions - Advanced captions settings - "Set custom captions" is not working (the edit captions window doesn’t open)
