+++ src/components/settings/settings.js
@@ -5,7 +5,7 @@ import {withText, Text} from 'preact-i18n';
 import {connect} from 'react-redux';
 import {bindActions} from 'utils';
 import {actions} from 'reducers/settings';
-import {AdvancedAudioDescToggle, AudioMenu, CaptionsMenu, QualityMenu, SmartContainer, SpeedMenu, HeightResolution} from 'components';
+import {AdvancedAudioDescToggle, AudioMenu, CaptionsMenu, QualityMenu, SmartContainer, SpeedMenu, HeightResolution, CVAAOverlay} from 'components';
 import {default as Icon, IconType, BadgeType} from '../icon';
 import {withPlayer} from '../player';
 import {withEventManager} from 'event/with-event-manager';
@@ -13,6 +13,7 @@ import {actions as overlayIconActions} from 'reducers/overlay-action';
 import {Tooltip} from 'components/tooltip';
 import {Button} from 'components/button';
 import {ButtonControl} from 'components/button-control';
+import {createPortal} from 'preact/compat';
 
 /**
  * mapping state to props
@@ -58,7 +59,7 @@ class Settings extends Component {
    * @memberof Settings
    */
   componentWillMount() {
-    this.setState({smartContainerOpen: false});
+    this.setState({smartContainerOpen: false, cvaaOverlay: false});
   }
 
   /**
@@ -81,6 +82,7 @@ class Settings extends Component {
    */
   handleClickOutside(e: any) {
     if (
+      !this.state.cvaaOverlay &&
       !this.props.isMobile &&
       !this.props.isSmallSize &&
       !!this._controlSettingsElement &&
@@ -100,6 +102,12 @@ class Settings extends Component {
   onControlButtonClick = (): void => {
     this.setState(prevState => {
       return {smartContainerOpen: !prevState.smartContainerOpen};
+    });
+  };
+
+  toggleCVAAOverlay = (): void => {
+    this.setState(prevState => {
+      return {cvaaOverlay: !prevState.cvaaOverlay};
     });
   };
 
@@ -150,6 +158,9 @@ class Settings extends Component {
     if (!(showAudioMenu || showCaptionsMenu || showQualityMenu || showSpeedMenu)) return undefined;
     if (props.isLive && props.videoTracks.length <= 1) return undefined;
     const buttonBadgeType: string = this.getButtonBadgeType() || '';
+
+    const targetId = document.getElementById(this.props.player.config.targetId) || document;
+    const portalSelector = `.overlay-portal`;
     return (
       <ButtonControl name={COMPONENT_NAME} ref={c => (c ? (this._controlSettingsElement = c) : undefined)}>
         <Tooltip label={props.buttonLabel}>
@@ -172,11 +183,12 @@ class Settings extends Component {
           <SmartContainer targetId={props.player.config.targetId} title={<Text id="settings.title" />} onClose={this.onControlButtonClick}>
             {showAdvancedAudioDescToggle && <AdvancedAudioDescToggle />}
             {showAudioMenu && <AudioMenu />}
-            {showCaptionsMenu && <CaptionsMenu />}
+            {showCaptionsMenu && <CaptionsMenu onAdvancedCaptionsClick={this.toggleCVAAOverlay} />}
             {showQualityMenu && <QualityMenu />}
             {showSpeedMenu && <SpeedMenu />}
           </SmartContainer>
         )}
+        {this.state.cvaaOverlay ? createPortal(<CVAAOverlay onClose={this.toggleCVAAOverlay} />, targetId.querySelector(portalSelector)) : <div />}
       </ButtonControl>
     );
   }
